refactor: what sir wants
