+++ graphql-api/db/bankcard-db.go
@@ -1,0 +1,8 @@
+package db
+
+
+/**
+	Supabase Connection string not client for postgre db
+	since sir wants us to use use pressly for migrations
+	dont use local postgredb since it will conflict our ports and stuffs
+**/
