+++ graphql-api/db/notification-db.go
@@ -1,0 +1,1 @@
+package db
