+++ graphql-api/services/account-service.go
@@ -1,3 +1,4 @@
-package helpers
+package services
 
+// Use this for resolvers business logic
 // Planning on putting helper functions here that can basically do CRUD to the DB
