+++ graphql-api/services/notification-service.go
@@ -1,0 +1,3 @@
+package services
+
+// Use this for resolvers business logic
